requirements.txt
streamlit
git+https://github.com/nmcassa/letterboxdpy.git
requests
python-dotenv
